saas-backend/requirements.txt
# PostgreSQL client for Python
psycopg2-binary
# ============================================================================
# Collective Access SaaS Backend - Dependencies
# Phase 3: Automated tenant provisioning and management
# ============================================================================

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pymysql>=1.0.3


# Kubernetes Client
kubernetes==29.0.0

# Stripe Integration
stripe==8.0.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.1.0

# HTTP Client
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
